\name{ARRmNormalization-package}
\alias{ARRmNormalization-package}
\alias{ARRmNormalization}
\docType{package}
\title{
ARRm normalization for Illumina methylation data
}
\description{	
Normalize Illumina methylation data from the Infinium HumanMethylation 450k assay with the Adaptive Robust Regression method. The normalization takes care of background intensity, dye bias, chip effects and spatial positions. The normalization can be applied to Beta values, M-values or other metrics as well. 
}
\author{
Jean-Philippe Fortin <jfortin@jhsph.edu>
Celia M.T. Greenwood <celia.greenwood@mcgill.ca>
Aurelie Labbe <aurelie.labbe@mcgill.ca>
}
\keyword{ package }

